+++ src/oops/SOLID/singleResponsibilityPrinciple/before/EmployeeDetailsPersistence.java
@@ -1,0 +1,5 @@
+package oops.SOLID.singleResponsibilityPrinciple.before;
+
+public interface EmployeeDetailsPersistence {
+    void save();
+}
